Begin refactoring work
