+++ jsconfig.json
@@ -3,8 +3,9 @@
     "baseUrl": "./",
     "paths": {
       "components/*": ["src/components/*"],
+      "helpers/*": ["src/helpers/*"],
       "libs/*": ["src/libs/*"],
-      "auth": ["src/auth"]
+      "auth": ["src/auth"],
     }
   }
 }
